requirements.txt
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pytest==7.4.3